+++ src/modules/keepOnlyCommand.ts
@@ -1,13 +1,19 @@
 import * as vscode from "vscode";
-import { NO_ACTIVE_EDITOR } from "../consts";
+import { COMMAND_CALL_ERROR, NO_ACTIVE_EDITOR } from "../consts";
 import { getSelectionLines, getSelectionsOrFullDocument, replaceSelectionsWithLines } from "../helpers/vsCodeHelpers";
 
-interface KeepDuplicatesOptions {
+interface KeepOnlyCommandOptions {
+	what: "duplicates" | "uniques";
 	onlyAdjacent: boolean;
 	caseSensitive: boolean;
 }
 
-export async function runKeepDuplicatesCommand(options: KeepDuplicatesOptions) {
+export async function runKeepOnlyCommand(options: KeepOnlyCommandOptions) {
+	if (options.what === "uniques" && options.onlyAdjacent === true) {
+		vscode.window.showErrorMessage(COMMAND_CALL_ERROR + " runKeepOnlyCommand, what: uniques, onlyAdjacent: true");
+		return;
+	}
+
 	const editor = vscode.window.activeTextEditor;
 	if (!editor) {
 		vscode.window.showWarningMessage(NO_ACTIVE_EDITOR);
@@ -21,7 +27,7 @@ export async function runKeepDuplicatesCommand(options: KeepDuplicatesOptions) {
 		matchingLinesBySelection.push([]);
 		const currentSelectionLines = matchingLinesBySelection[matchingLinesBySelection.length - 1];
 
-		let lineCounter: { [index: string]: number } = {};
+		let lineCounter: Map<string, number> = new Map();
 		let lastLine: string | null = null;
 
 		for (const lineContent of getSelectionLines(editor, selection)) {
@@ -30,20 +36,32 @@ export async function runKeepDuplicatesCommand(options: KeepDuplicatesOptions) {
 				: lineContent.toLocaleLowerCase();
 
 			if (options.onlyAdjacent && normalizedLineContent !== lastLine) {
-				lineCounter = {};
+				lineCounter.clear();
 			}
 
-			if (lineCounter[normalizedLineContent]) {
-				lineCounter[normalizedLineContent]++;
+			if (lineCounter.has(normalizedLineContent)) {
+				lineCounter.set(normalizedLineContent, lineCounter.get(normalizedLineContent)! + 1);
 
-				if (lineCounter[normalizedLineContent] === 2) {
+				if (options.what === "duplicates" && lineCounter.get(normalizedLineContent) === 2) {
 					currentSelectionLines.push(lineContent);
 				}
 			} else {
-				lineCounter[normalizedLineContent] = 1;
+				lineCounter.set(normalizedLineContent, 1);
 			}
 
 			lastLine = normalizedLineContent;
+		}
+
+		if (options.what === "uniques") {
+			for (const lineContent of getSelectionLines(editor, selection)) {
+				const normalizedLineContent = options.caseSensitive
+					? lineContent
+					: lineContent.toLocaleLowerCase();
+
+				if (lineCounter.get(normalizedLineContent) === 1) {
+					currentSelectionLines.push(lineContent);
+				}
+			}
 		}
 	}