+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## **1.43.0** (2023. 09. 16.)
 * Pad counters in the output of _Count line occurrences_ and _Count adjacent duplicates_ commands, so the result can be sorted using standard sorting commands.
+* Add _Keep only unique lines_ and _Keep only unique lines (case insensitive)_ commands (#72)
 
 ## **1.42.0** (2023. 09. 09.)
 * Add _Remove duplicated lines (case insensitive)_ command